Fix test

+++ tests/integration_tests/datasets/api_tests.py
@@ -153,7 +153,8 @@ class TestDatasetApi(SupersetTestCase):
 
             # rollback changes
             for dataset in datasets:
-                db.session.delete(dataset)
+                if not hasattr(dataset, "_deleted"):
+                    db.session.delete(dataset)
             db.session.commit()
 
     @staticmethod
@@ -1711,6 +1712,8 @@ class TestDatasetApi(SupersetTestCase):
         assert rv.status_code == 200
         expected_response = {"message": f"Deleted {len(datasets)} datasets"}
         assert data == expected_response
+        for dataset in datasets:
+            setattr(dataset, "_deleted", True)
         datasets = (
             db.session.query(SqlaTable)
             .filter(SqlaTable.table_name.in_(self.fixture_tables_names))
